+++ CodeRed/Shared/Enum/ResourceLayout.hpp
@@ -6,7 +6,6 @@ namespace CodeRed {
 
 	enum class ResourceLayout : UInt32
 	{
-		Undefined,
 		GeneralRead,
 		RenderTarget,
 		DepthStencil,